@@ -4,41 +4,40 @@ using Cqrs.Template.Domain.Exceptions;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Cqrs.Template.Api.Controllers
+namespace Cqrs.Template.Api.Controllers;
+
+[Route("your-project-name/[controller]/v{version:apiVersion}")]
+[ServiceFilter(typeof(GlobalExceptionFilterAttribute))]
+public class BaseController : Controller
 {
-	[Route("your-project-name/[controller]/v{version:apiVersion}")]
-	[ServiceFilter(typeof(GlobalExceptionFilterAttribute))]
-	public class BaseController : Controller
+	private readonly ExceptionNotificationHandler _notifications;
+
+	protected IEnumerable<ExceptionNotification> Notifications => _notifications.GetNotifications();
+
+	protected BaseController(INotificationHandler<ExceptionNotification> notifications)
 	{
-		private readonly ExceptionNotificationHandler _notifications;
+		_notifications = (ExceptionNotificationHandler) notifications;
+	}
 
-		protected IEnumerable<ExceptionNotification> Notifications => _notifications.GetNotifications();
+	protected bool IsValidOperation()
+	{
+		return (!_notifications.HasNotifications());
+	}
 
-		protected BaseController(INotificationHandler<ExceptionNotification> notifications)
+	protected new IActionResult Response(IActionResult action)
+	{
+		if (IsValidOperation())
 		{
-			_notifications = (ExceptionNotificationHandler) notifications;
+			return action;
 		}
 
-		protected bool IsValidOperation()
-		{
-			return (!_notifications.HasNotifications());
-		}
-
-		protected new IActionResult Response(IActionResult action)
-		{
-			if (IsValidOperation())
+		return BadRequest
+		(
+			new
 			{
-				return action;
+				success = false,
+				errors = _notifications.GetNotifications()
 			}
-
-			return BadRequest
-			(
-				new
-				{
-					success = false,
-					errors = _notifications.GetNotifications()
-				}
-			);
-		}
+		);
 	}
 }
