@@ -1,17 +1,44 @@
-using Microsoft.AspNetCore.Hosting;
+using Cqrs.Template.Api.Filters;
+using Cqrs.Template.Application.CommandHandlers;
+using Cqrs.Template.Infra.CrossCutting.IoC.Configurations;
+using MediatR;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Mvc.ApiExplorer;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Cqrs.Template.Api
-{
-	public class Program
-	{
-		public static void Main(string[] args)
-		{
-			CreateHostBuilder(args).Build().Run();
-		}
+var builder = WebApplication.CreateBuilder(args);
 
-		public static IHostBuilder CreateHostBuilder(string[] args) =>
-			Host.CreateDefaultBuilder(args)
-				.ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); });
-	}
+builder.Services.AddApiVersioning();
+builder.Services.AddVersionedApiExplorer();
+builder.Services.AddSwaggerSetup();
+builder.Services.AddAutoMapper();
+builder.Services.AddDependencyInjectionSetup();
+builder.Services.AddMediatR(typeof(CommandHandler));
+builder.Services.AddScoped<GlobalExceptionFilterAttribute>();
+builder.Services.AddDatabaseSetup();
+builder.Services.AddControllers();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.UseDeveloperExceptionPage();
 }
+
+app.UseCors(builder =>
+{
+    builder.WithOrigins("*");
+    builder.AllowAnyOrigin();
+    builder.AllowAnyMethod();
+    builder.AllowAnyHeader();
+});
+
+app.UseRouting();
+
+var apiVersionDescriptionProvider = app.Services.GetRequiredService<IApiVersionDescriptionProvider>();
+app.UseSwaggerSetup(apiVersionDescriptionProvider);
+
+app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
+
+app.Run();
