@@ -2,15 +2,14 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 
-namespace Cqrs.Template.Application.EventHandlers
-{
-	public abstract class EventHandler<T> : INotificationHandler<T> where T : INotification
-	{
-		protected EventHandler()
-		{
-			
-		}
+namespace Cqrs.Template.Application.EventHandlers;
 
-		public abstract Task Handle(T notification, CancellationToken cancellationToken);
+public abstract class EventHandler<T> : INotificationHandler<T> where T : INotification
+{
+	protected EventHandler()
+	{
+			
 	}
+
+	public abstract Task Handle(T notification, CancellationToken cancellationToken);
 }
