@@ -1,7 +1,6 @@
-namespace Cqrs.Template.Domain.SeedWork
+namespace Cqrs.Template.Domain.SeedWork;
+
+public interface IRepository<TEntity> where TEntity : IAggregateRoot
 {
-	public interface IRepository<TEntity> where TEntity : IAggregateRoot
-	{
-		void Add(TEntity obj);
-	}
+	void Add(TEntity obj);
 }
