@@ -2,15 +2,14 @@ using System;
 using Cqrs.Template.Infra.Data.Context;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Cqrs.Template.Infra.CrossCutting.IoC.Configurations
-{
-	public static class DatabaseSetup
-	{
-		public static void AddDatabaseSetup(this IServiceCollection services)
-		{
-			if (services == null) throw new ArgumentNullException(nameof(services));
+namespace Cqrs.Template.Infra.CrossCutting.IoC.Configurations;
 
-			services.AddDbContext<ApplicationDbContext>(ServiceLifetime.Scoped);
-		}
+public static class DatabaseSetup
+{
+	public static void AddDatabaseSetup(this IServiceCollection services)
+	{
+		if (services == null) throw new ArgumentNullException(nameof(services));
+
+		services.AddDbContext<ApplicationDbContext>(ServiceLifetime.Scoped);
 	}
 }
