@@ -1,15 +1,14 @@
 using System;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Cqrs.Template.Infra.CrossCutting.IoC.Configurations
-{
-	public static class DependencyInjectionSetup
-	{
-		public static void AddDependencyInjectionSetup(this IServiceCollection services)
-		{
-			if (services == null) throw new ArgumentNullException(nameof(services));
+namespace Cqrs.Template.Infra.CrossCutting.IoC.Configurations;
 
-			NativeInjectorBootstrapper.RegisterServices(services);
-		}
+public static class DependencyInjectionSetup
+{
+	public static void AddDependencyInjectionSetup(this IServiceCollection services)
+	{
+		if (services == null) throw new ArgumentNullException(nameof(services));
+
+		NativeInjectorBootstrapper.RegisterServices(services);
 	}
 }
