@@ -4,27 +4,26 @@ using Cqrs.Template.Domain.SeedWork;
 using Cqrs.Template.Infra.Data.Context;
 using MediatR;
 
-namespace Cqrs.Template.Infra.Data
+namespace Cqrs.Template.Infra.Data;
+
+static class MediatorExtension
 {
-	static class MediatorExtension
+	public static async Task DispatchDomainEventsAsync(this IMediator mediator, ApplicationDbContext ctx)
 	{
-		public static async Task DispatchDomainEventsAsync(this IMediator mediator, ApplicationDbContext ctx)
+		var domainEntities = ctx.ChangeTracker
+			.Entries<Entity>()
+			.Where(x => x.Entity.DomainEvents != null && x.Entity.DomainEvents.Any());
+
+		var domainEvents = domainEntities
+			.SelectMany(x => x.Entity.DomainEvents)
+			.ToList();
+
+		domainEntities.ToList()
+			.ForEach(entity => entity.Entity.ClearDomainEvent());
+
+		foreach (var domainEvent in domainEvents)
 		{
-			var domainEntities = ctx.ChangeTracker
-				.Entries<Entity>()
-				.Where(x => x.Entity.DomainEvents != null && x.Entity.DomainEvents.Any());
-
-			var domainEvents = domainEntities
-				.SelectMany(x => x.Entity.DomainEvents)
-				.ToList();
-
-			domainEntities.ToList()
-				.ForEach(entity => entity.Entity.ClearDomainEvent());
-
-			foreach (var domainEvent in domainEvents)
-			{
-				await mediator.Publish(domainEvent);
-			}
+			await mediator.Publish(domainEvent);
 		}
 	}
 }
